Add region support for scribble text
3 new `text` properties
regions: true/false to enable region support (colors and/or events)
region_color: blend color when region is hovered. if set, `regions` will be automatically set to true
region_blend: defaults to 1

1 new `text` event
on_region_hover_changed: this will be triggered when the hovered region changes, the parameter is a struct with the following properties:
source: the YUI element that triggered the event
scribble_element: the scribble_element that the region is defined on
region: the name of the region that is currently hovered. will be `undefined` when the mouse stops hovering over a region